Move API to Images Folder, Docker Setup, and Class Restructure
Moved the API codebase to the "images" folder for better organization.

Set up Docker with a Dockerfile for the API service and a docker-compose.yml file.

Updated the docker-compose.yml file to use the new directory structure.

Created classes (Game.js and User.js) within the "classes" folder under the API.

This commit restructures the project to enhance clarity and incorporates Docker for containerization. The new directory structure improves the organization of code, and the classes are now placed in a dedicated folder. The Docker setup allows for easier deployment and testing of the API with a PostgreSQL database.